% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{npar-package}
\alias{npar}
\alias{npar-package}
\title{Functions for nonparametric group comparisons.}
\description{
npar provides tools for calculating and visualizing
nonparametric differences among groups.
}

