% Generated by roxygen2 (4.0.2): do not edit by hand
\name{oneway}
\alias{oneway}
\title{Nonparametric group comparisons}
\usage{
oneway(formula, data, exact = FALSE, sort = TRUE, method = c("holm",
  "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"))
}
\arguments{
\item{formula}{an object of class formula, relating the dependent variable
to the grouping variable.}

\item{data}{a data frame containing the variables in the model.}

\item{exact}{logical. If \code{TRUE}, calculate exact Wilcoxon tests.}

\item{sort}{logical. If \code{TRUE}, sort groups by median dependent variable values.}

\item{method}{method for correcting p-values for multiple comparisons.}
}
\value{
a list with 7 elements:
\item{CALL}{function call}
\item{data}{data frame containg the depending and grouping variable}
\item{sumstats}{data frame with descriptive statistics by group}
\item{kw}{results of the Kruskal-Wallis test}
\item{method}{method used to adjust p-values}
\item{wmc}{data frame containing the multiple comparisons}
\item{vnames}{variable names}
}
\description{
\code{oneway} computes nonparametric group comparisons, including an
omnibus test and post-hoc pairwise group comparisons.
}
\details{
This function computes an omnibus Kruskal-Wallis test that the
groups are equal, followed by all pairwise comparisons using
Wilcoxon Rank Sum tests. Exact Wilcoxon tests can be requested if
there are no ties on the dependent variable. The p-values are
adjusted for mulitple comparisons using the \code{\link{p.adjust}}
function.
}
\examples{
results <- oneway(hlef ~ region, life)
summary(results)
plot(results, col="lightblue", main="Multiple Comparisons",
     xlab="US Region", ylab="Healthy Life Expectancy at Age 65")
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}

