% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot nonparametric group comparisons}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}.}

\item{...}{additional arguments passed to the
\code{\link{boxplot}} function.}
}
\description{
\code{plot.oneway} plots nonparametric group comparisons.
}
\details{
This function plots nonparametric group comparisons
created by the \code{\link{oneway}} function using
annotated side by side boxplots. Medians and
sample sizes are placed at the top of the chart.
The overall median is represented by a horizontal
dashed line.
}
\examples{
results <- oneway(hlef ~ region, life)
plot(results, col="lightblue", main="Multiple Comparisons",
     xlab="US Region", ylab="Healthy Life Expectancy at Age 65")
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}

