% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{打印多重比较的结果}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{一个 \code{oneway}类型的变量}

\item{...}{要传输给函数的额外的变量}
}
\value{
静默返回输入的变量
}
\description{
\code{print.oneway} 打印多重组间比较的结果
}
\details{
这个函数打印出用 \code{\link{oneway}} 函数所创建的Wilcoxon成对多重比较的结果
}
\examples{
results <- oneway(hlef ~ region, life)
print(results)
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}
