% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarize oneway nonparametric analyses}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{an object of class \code{oneway}.}

\item{...}{additional parameters.}
}
\value{
the input object is returned silently.
}
\description{
\code{summary.oneway} summarizes the results of a oneway
nonparametric analysis.
}
\details{
This function prints a summary of analyses produced by
the \code{\link{oneway}} function. This includes descriptive
statistics by group, an omnibus Kruskal-Wallis test, and
Wilcoxon pairwise multiple comparisons.
}
\examples{
results <- oneway(hlef ~ region, life)
summary(results)
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}

