% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npar.R
\docType{package}
\name{npar-package}
\alias{npar-package}
\alias{npar}
\title{Functions for nonparametric group comparisons.}
\description{
npar provides tools for calculating and visualizing
nonparametric differences among groups.
}
