% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{非参组间比较}
\usage{
oneway(formula, data, exact = FALSE, sort = TRUE, method = c("holm",
  "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"))
}
\arguments{
\item{formula}{一个formula对象。用于表示因变量和分组变量之间的关系}

\item{data}{一个包含了模型里变量的数据框}

\item{exact}{logical变量。如\code{TRUE}，计算精确的Wilcoxon检验}

\item{sort}{logical变量，如果\code{TRUE}，用因变量中位数来对组别进行排序}

\item{method}{用于调整多重比较的p值的方法}
}
\value{
一个有7个元素的列表
\item{CALL}{函数调用}
\item{data}{包含因变量和组间变量的数据框}
\item{sumstats}{包含每组的描述性统计量的数据框}
\item{kw}{K-W检验的结果}
\item{method}{用于调整p值的方法}
\item{wmc}{包含多重比较的数据框}
\item{vnames}{变量名}
}
\description{
\code{oneway} 计算非参组间比较，包括综合检验和事后成对组间比较
}
\details{
这个函数计算了一个综合Kruskal-Wallis检验，用于检验组别是否相等，接着使用
Wilcoxon秩和检验来进行成对比较。如果因变量之间没有相互依赖的话，可以计算精确
的Wilcoxon检验。使用\code{\link{p.adjust}}来对多重比较所得到的p值进行调整
}
\examples{
results <- oneway(hlef ~ region, life)
summary(results)
plot(results, col="lightblue", main="Multiple Comparisons",
     xlab="US Region", ylab="Healthy Life Expectancy at Age 65")
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}
