% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{对非参组间比较的结果进行绘图}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{一个\code{oneway}类型的对象}

\item{...}{被传递给
\code{\link{boxplot}} 的额外参数}
}
\description{
\code{plot.oneway} 对非参组间比较的结果进行绘图
}
\details{
这个函数使用标记了的并排箱线图对\code{\link{oneway}}函数所生成的非参组间比较
结果进行绘图。中位数和样本量被放置在图的上方、总体中位数用一条虚线进行表示
}
\examples{
results <- oneway(hlef ~ region, life)
plot(results, col="lightblue", main="Multiple Comparisons",
     xlab="US Region", ylab="Healthy Life Expectancy at Age 65")
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}
