% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.oneway}
\alias{print.oneway}
\title{Print multiple comparisons}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}.}

\item{...}{additional arguments passed to the function.}
}
\value{
the input object is returned silently.
}
\description{
\code{print.oneway} prints pairwise group comparisons.
}
\details{
This function prints Wilcoxon pairwise multiple comparisons created
by the \code{\link{oneway}} function.
}
\examples{
results <- oneway(hlef ~ region, life)
print(results)
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}

