% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{汇总单因子非参分析的结果}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{一个\code{oneway}类型的对象}

\item{...}{额外的参数}
}
\value{
静默返回输入的对象
}
\description{
\code{summary.oneway} 汇总了单因子非参分析的结果
}
\details{
这个函数对\code{\link{oneway}}函数所分析的结果进行汇总并打印。这包括了每一组的
描述性统计量，一个综合的Kruskal-Wallis检验的结果，以及一个Wilcoxon成对多重比较
的结果
}
\examples{
results <- oneway(hlef ~ region, life)
summary(results)
}
\author{
Rob Kabacoff <rkabacoff@statmethods.net>
}
